Rename tool_menu -> ToolMenu, etc
