+++ dllmain/GameFlags.h
