+++ dllmain/dllmain.cpp
@@ -9,10 +9,10 @@
 #include "Sofdec.h"
 #include "FilterXXFixes.h"
 #include "UltraWideFix.h"
-#include "tool_menu.h"
+#include "ToolMenu.h"
 #include "Settings.h"
 #include "ConsoleWnd.h"
-#include "qtefixes.h"
+#include "QTEFixes.h"
 #include "60fpsFixes.h"
 #include "KeyboardMouseTweaks.h"
 #include "ExceptionHandler.h"
